Fixed some g++ warnings about int/size_t comparisons.  Also fixed a bug in ConfigFile converting from {a,b,c} syntax to a vector<string>.
git-svn-id: https://desweb.cosmology.illinois.edu/svn/desdm/devel/wl/trunk@5945 2422e80f-9519-0410-a819-8e21e7c86b6d

+++ src/BVec.h
@@ -55,7 +55,7 @@ public :
     int getOrder() const { return _order; }
     double getSigma() const { return _sigma; }
     const DVector& getValues() const { return _b; }
-    int size() const { return _b.size(); }
+    size_t size() const { return _b.size(); }
 
     void setSigma(double sigma) { _sigma = sigma; }
